fix: firmware

+++ internal/dto/firmware_index_response.go
@@ -1,6 +1,6 @@
 package dto
 
 type FirmwareIndexResponse struct {
-	Version int    `json:"version"`
-	Source  string `json:"source"`
+	Version float32 `json:"version"`
+	Source  string  `json:"source"`
 }
